requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
PyMySQL==1.0.2
Werkzeug==2.0.3
python-dotenv==0.20.0
SQLAlchemy==1.4.46
